+++ README.md
@@ -82,7 +82,7 @@ MATH_PIを使ってラジアンを60分法に変換します。
 
 doxygen追加済み  
 
-## Chipher.h
+## Cipher.h
 暗号を扱う関数が入っています。  
 全ての関数の引数にある**modeという引数は暗号化を行うか復号化を行うかを決めるための引数で、'm'で暗号化，'r'で復号化となります。**  
 以下の関数が入っています。  
@@ -637,20 +637,20 @@ doxygen追加済み
 
 doxygen追加済み
 
-## Chipher.hpp
-暗号を扱うChipherクラスが入っています。  
+## Cipher.hpp
+暗号を扱うCipherクラスが入っています。  
 どの関数でも**モードは"m"で暗号化、"r"で解読です**。
 以下の関数が入っています。  
 * string Cipher::ceasar(string モード,string 平文or暗号文,int シフトする数)  
 :シーザー暗号を扱います。
 * string Cipher.vigenere(string モード，string 平文or暗号文，string 鍵)
 :ヴィジュネル暗号を扱います。
-* string Chipher::substitution(string モード，string 平文or暗号文，string 鍵)  
+* string Cipher::substitution(string モード，string 平文or暗号文，string 鍵)  
 :単一換字式暗号を扱います。鍵にa~zに対応した文字列を渡します。
-* string Chipher::polybius_square(string モード，string 平文or暗号文)  
+* string Cipher::polybius_square(string モード，string 平文or暗号文)  
 :ポリュビオスの暗号表(5\*5)を扱います。  
 5\*5の方式ではjとiの暗号が同じ結果になります。そのため、復号結果ではiはiのままjをiとします。(平文ではa-z以外の文字(空白も含む)は除いてください。暗号文では0と1以外の文字(空白も含む)は除いてください。)  
-* string Chipher::scytale(string モード，string 平文or暗号文)
+* string Cipher::scytale(string モード，string 平文or暗号文)
 スキュタレー暗号を扱います。  
 5列に分けて暗号化・復号化を行います。  
 言語の仕様上変数の値を配列の要素数の宣言に使用できなかったので、500個(499文字)までに対応させました。  
@@ -724,16 +724,16 @@ Javadoc追加済み
 ## Cipher.java
 暗号を扱うクラスが入っています。  
 以下の関数が入っています。  
-* String Chipher.ceasar(String モード,String 平文or暗号文,int シフトする数)  
+* String Cipher.ceasar(String モード,String 平文or暗号文,int シフトする数)  
 :シーザー暗号を扱います。**モードは"m"で暗号化、"r"で解読です(他関数でも同じです)**。
-* String Chipher.vigenere(String モード，String 平文or暗号文，String 鍵)  
+* String Cipher.vigenere(String モード，String 平文or暗号文，String 鍵)  
 :ヴィジュネル暗号を扱います。
-* String Chipher.substitution(String モード，String 平文or暗号文，String 鍵)  
+* String Cipher.substitution(String モード，String 平文or暗号文，String 鍵)  
 :単一換字式暗号を扱います。鍵にa~zに対応した文字列を渡します。  
-* String Chipher.polybius_square(String モード，String 平文or暗号文)  
+* String Cipher.polybius_square(String モード，String 平文or暗号文)  
 :ポリュビオスの暗号表(5\*5)を扱います。  
 5\*5の方式ではjとiの暗号が同じ結果になります。そのため、復号結果ではiはiのままjをiとします。(平文ではa-z以外の文字(空白も含む)は除いてください。暗号文では0と1以外の文字(空白も含む)は除いてください。)  
-* string Chipher.scytale(string モード，string 平文or暗号文)
+* string Cipher.scytale(string モード，string 平文or暗号文)
 スキュタレー暗号を扱います。  
 5列に分けて暗号化・復号化を行います。  
 
@@ -881,12 +881,12 @@ docstring追加済み
 :シーザー暗号を扱います。**モードは"m"で暗号化、"r"で解読です(他関数でも同じです)**。  
 * Cipher.vigenere(モード,平文or暗号文,鍵) 
 :ヴィジュネル暗号を扱います。
-* Chipher.substitution(モード,平文or暗号文,鍵)  
+* Cipher.substitution(モード,平文or暗号文,鍵)  
 :単一換字式暗号を扱います。鍵にa~zに対応した文字列を渡します。  
-* Chipher.polybius_square(モード,平文or暗号文)  
+* Cipher.polybius_square(モード,平文or暗号文)  
 :ポリュビオスの暗号表(5\*5)を扱います。  
 5\*5の方式ではjとiの暗号が同じ結果になります。そのため、復号結果ではiはiのままjをiとします。(平文ではa-z以外の文字(空白も含む)は除いてください。暗号文では0と1以外の文字(空白も含む)は除いてください。)  
-* Chipher.scytale(モード,平文or暗号文)
+* Cipher.scytale(モード,平文or暗号文)
 :スキュタレー暗号を扱います。  
 ５列に分けて暗号化します。
